updated buy policy to matl buy policy with s, S inventory management. added initial test
